Print all of the restaurants' names
Using next_page_token to get the rest of the results suck, google makes
you wait a few seconds before you send another request (remove the 2
seconds sleep and the code breaks when you run it - give it a try). Next
up on the list would probably be storing this data in an arraylist or
something (if you want to tackle that, it'll be
ArrayList<PlaceSearch.PlaceData> - placedata is an inner class in
placesearch that holds the actual restaurant info. because it's defined
inside of another class, you can't refer to it like this:
ArrayList<PlaceData> (if you try typing that into netbeans it yells at
you))